Refactor UploadHandler
